[Imp] Update r8brain to r72.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4658 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ include/r8brain/CDSPBlockConvolver.h
@@ -294,7 +294,7 @@ public:
 				ilu = InputLen;
 			}
 
-			const int pil = PrevInputLen * sizeof( double );
+			const int pil = (int) ( PrevInputLen * sizeof( double ));
 			memcpy( &CurInput[ ilu ], PrevInput, pil );
 			memcpy( PrevInput, &CurInput[ ilu - PrevInputLen ], pil );
